PEER-219: Refactor attempts files
