+++ frontend/src/components/blocks/interview/question-attempts/index.ts
@@ -1,0 +1,1 @@
+export * from './main';
