+++ frontend/src/components/blocks/interview/question-attempts/table.tsx
